# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: Enterprise Platform API - LMS, Tasks, Files, Automations
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email/password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /courses:
    get:
      operationId: getCourses
      tags: [courses]
      summary: List all courses
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, published, archived]
      responses:
        "200":
          description: List of courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
    post:
      operationId: createCourse
      tags: [courses]
      summary: Create a course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCourseRequest"
      responses:
        "201":
          description: Course created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /courses/{id}:
    get:
      operationId: getCourse
      tags: [courses]
      summary: Get course by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Course with modules and lessons
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      operationId: updateCourse
      tags: [courses]
      summary: Update course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCourseRequest"
      responses:
        "200":
          description: Course updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
    delete:
      operationId: deleteCourse
      tags: [courses]
      summary: Delete course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /courses/{id}/enroll:
    post:
      operationId: enrollUserInCourse
      tags: [courses]
      summary: Enroll a user in a course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: integer
              required: [userId]
      responses:
        "200":
          description: Enrolled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /courses/{id}/progress:
    get:
      operationId: getCourseProgress
      tags: [courses]
      summary: Get progress for all users in a course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Progress list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Progress"
  /courses/{courseId}/modules:
    get:
      operationId: getCourseModules
      tags: [courses]
      summary: Get modules for a course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Modules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Module"
    post:
      operationId: createModule
      tags: [courses]
      summary: Create a module
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateModuleRequest"
      responses:
        "201":
          description: Module created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Module"
...
  /files:
    get:
      operationId: getFiles
      tags: [files]
      summary: List files
      parameters:
        - name: folder
          in: query
          schema:
            type: string
        - name: userId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FileRecord"
    post:
      operationId: uploadFile
      tags: [files]
      summary: Upload file metadata
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFileRequest"
      responses:
        "201":
          description: File record created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileRecord"
  /files/{id}:
    get:
      operationId: getFile
      tags: [files]
      summary: Get file by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: File
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileRecord"
    delete:
      operationId: deleteFile
      tags: [files]
      summary: Delete file
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /automations:
    get:
      operationId: getAutomations
      tags: [automations]
      summary: List automations
      responses:
        "200":
          description: Automations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Automation"
    post:
      operationId: createAutomation
      tags: [automations]
      summary: Create automation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAutomationRequest"
      responses:
        "201":
          description: Automation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Automation"
  /automations/{id}:
    put:
      operationId: updateAutomation
      tags: [automations]
      summary: Update automation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAutomationRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Automation"
    delete:
      operationId: deleteAutomation
      tags: [automations]
      summary: Delete automation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, manager, employee]
        department:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, email, name, role, isActive, createdAt]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required: [user, token]
...
    Course:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        thumbnailUrl:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, published, archived]
        duration:
          type: integer
          nullable: true
          description: Duration in minutes
        createdById:
          type: integer
        enrolledCount:
          type: integer
        completionRate:
          type: number
          format: float
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, title, status, createdById, enrolledCount, completionRate, createdAt, updatedAt]
    CourseDetail:
      allOf:
        - $ref: "#/components/schemas/Course"
        - type: object
          properties:
            modules:
              type: array
              items:
                $ref: "#/components/schemas/ModuleWithLessons"
    Module:
      type: object
      properties:
        id:
          type: integer
        courseId:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        order:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, courseId, title, order, createdAt]
    ModuleWithLessons:
      allOf:
        - $ref: "#/components/schemas/Module"
        - type: object
          properties:
            lessons:
              type: array
              items:
                $ref: "#/components/schemas/Lesson"
    Lesson:
      type: object
      properties:
        id:
          type: integer
        moduleId:
          type: integer
        title:
          type: string
        content:
          type: string
          nullable: true
        videoUrl:
          type: string
          nullable: true
        duration:
          type: integer
          nullable: true
        order:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, moduleId, title, order, createdAt]
    Progress:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        courseId:
          type: integer
        lessonId:
          type: integer
          nullable: true
        status:
          type: string
          enum: [not_started, in_progress, completed]
        completedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
          nullable: true
      required: [id, userId, courseId, status, createdAt]
    CreateCourseRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        thumbnailUrl:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, published, archived]
        duration:
          type: integer
          nullable: true
        createdById:
          type: integer
      required: [title, status, createdById]
    UpdateCourseRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        thumbnailUrl:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, published, archived]
        duration:
          type: integer
          nullable: true
    CreateModuleRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        order:
          type: integer
      required: [title, order]
...
    FileRecord:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        originalName:
          type: string
        mimeType:
          type: string
        size:
          type: integer
        folder:
          type: string
          nullable: true
        url:
          type: string
        uploadedById:
          type: integer
        isPublic:
          type: boolean
        createdAt:
          type: string
          format: date-time
        uploadedBy:
          $ref: "#/components/schemas/User"
          nullable: true
      required: [id, name, originalName, mimeType, size, url, uploadedById, isPublic, createdAt]
    CreateFileRequest:
      type: object
      properties:
        name:
          type: string
        originalName:
          type: string
        mimeType:
          type: string
        size:
          type: integer
        folder:
          type: string
          nullable: true
        url:
          type: string
        uploadedById:
          type: integer
        isPublic:
          type: boolean
      required: [name, originalName, mimeType, size, url, uploadedById, isPublic]
    Automation:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        trigger:
          type: string
          enum: [course_completed, task_completed, file_uploaded, user_enrolled, task_created]
        condition:
          type: object
          nullable: true
          additionalProperties: true
        action:
          type: string
          enum: [assign_task, notify_admin, send_notification, create_task, enroll_course]
        actionConfig:
          type: object
          additionalProperties: true
        isActive:
          type: boolean
        runCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, name, trigger, action, actionConfig, isActive, runCount, createdAt]
    CreateAutomationRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        trigger:
          type: string
          enum: [course_completed, task_completed, file_uploaded, user_enrolled, task_created]
        condition:
          type: object
          nullable: true
          additionalProperties: true
        action:
          type: string
          enum: [assign_task, notify_admin, send_notification, create_task, enroll_course]
        actionConfig:
          type: object
          additionalProperties: true
        isActive:
          type: boolean
      required: [name, trigger, action, actionConfig, isActive]
    UpdateAutomationRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        trigger:
          type: string
          enum: [course_completed, task_completed, file_uploaded, user_enrolled, task_created]
        condition:
          type: object
          nullable: true
          additionalProperties: true
        action:
          type: string
          enum: [assign_task, notify_admin, send_notification, create_task, enroll_course]
        actionConfig:
          type: object
          additionalProperties: true
        isActive:
          type: boolean
...