# lib/api-zod/src/generated/types/automation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Enterprise Platform API - LMS, Tasks, Files, Automations
 * OpenAPI spec version: 1.0.0
 */
import type { AutomationAction } from "./automationAction";
import type { AutomationActionConfig } from "./automationActionConfig";
import type { AutomationCondition } from "./automationCondition";
import type { AutomationTrigger } from "./automationTrigger";

export interface Automation {
  id: number;
  name: string;
  description?: string | null;
  trigger: AutomationTrigger;
  condition?: AutomationCondition;
  action: AutomationAction;
  actionConfig: AutomationActionConfig;
  isActive: boolean;
  runCount: number;
  createdAt: Date;
}