# lib/api-zod/src/generated/types/module.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Enterprise Platform API - LMS, Tasks, Files, Automations
 * OpenAPI spec version: 1.0.0
 */

export interface Module {
  id: number;
  courseId: number;
  title: string;
  description?: string | null;
  order: number;
  createdAt: Date;
}