# lib/api-zod/src/generated/types/updateCourseRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Enterprise Platform API - LMS, Tasks, Files, Automations
 * OpenAPI spec version: 1.0.0
 */
import type { UpdateCourseRequestStatus } from "./updateCourseRequestStatus";

export interface UpdateCourseRequest {
  title?: string;
  description?: string | null;
  thumbnailUrl?: string | null;
  status?: UpdateCourseRequestStatus;
  duration?: number | null;
}