# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Enterprise Platform API - LMS, Tasks, Files, Automations
 * OpenAPI spec version: 1.0.0
 */
import type { UserRole } from "./userRole";

export interface User {
  id: number;
  email: string;
  name: string;
  role: UserRole;
  department?: string | null;
  avatarUrl?: string | null;
  isActive: boolean;
  createdAt: Date;
}